backend/pyproject.toml
[tool.poetry]
name = "pofolpark-backend"
version = "0.1.0"
description = ""
authors = ["BORB <qzsec299@naver.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.7"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
openai = "^0.27.6"
django-cors-headers = "^4.0.0"
dj-database-url = "^2.0.0"
psycopg2-binary = "^2.9.6"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
gunicorn = "^20.1.0"
django-crontab = "^0.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
